Corrected debug prints in some commands
NOTE: LOG recording not yet implementing reset protection.

+++ Commands/CommandDeleteFilter.cpp
@@ -4,6 +4,7 @@
 #include "Debug/DebugMacros.h"
 #include "Log/LogServer.h"
 #include "Commands/CommandDescriptor.h"
+#include "System/Reset.h"
 
 CommandDeleteFilter::CommandDeleteFilter(QSharedPointer<LogServer> logServer, QObject *parent,
 						  void (*preprocessValidCommand)(CommandDescriptor &descriptor)) :
@@ -18,19 +19,25 @@ void CommandDeleteFilter::onCommandReceived(CommandDescriptor descriptor)
 	if (descriptor.value(QStringLiteral(COMMAND_NAME)) == QStringLiteral(COMMAND_DELETE_VALUE)) {
 			// --- DELETE COMMAND ---
 			_preprocessValidCommand(descriptor);
+
+			if(!System::startResetSensitiveProcess(RESET_PRIORITY_BULK)) {
+				DP_RESET_ERROR("NOT EXECUTING DELETE COMMAND: Reset protection in progress.");
+				return;
+			}
+
 			QString source(_logServer->pathname(descriptor.value(QStringLiteral(COMMAND_PARAMETER_SOURCE_FILE_NAME))));
 
 			if(source.isEmpty()) {
-				D_P("\tSource parameter missing or invalid:" << source << "\n\tAborting...");
+				DP_CMD_LOG_DELETE_ERROR("\tSource parameter missing or invalid:" << source << "\n\tAborting...");
 				return;
 			}
 
 			QSharedPointer<LogMaintenanceLocker> lock(_logServer->fileMaintenanceLocker());
 
 			if(QFile::remove(source))
-				D_P("\tRemoved:" << source);
+				DP_CMD_LOG_DELETE("\tRemoved:" << source);
 			else
-				D_P("\tRemoval FAILED !!! File may remain on disk:" << source);
+				DP_CMD_LOG_DELETE_ERROR("\tRemoval FAILED !!! File may remain on disk:" << source);
 	}
 
 }
